Correct error in IBIA CBEFF organization ID.
Participant informed NIST that IBIA notified participant of an error when
issuing their CBEFF biometric organization identifier. This created an issue for
MINEX III, since the participant used the erroneous CBEFF ID in their submission
name. Because the submission is not compliant and will not be run again, we are
simply changing the names of the result files.